hate it here fr run the right one... (fixes command to work with new data set)
