+++ Items/SpriteItemAnimated.cs
@@ -12,7 +12,7 @@ namespace LegendOfZelda
         List<Rectangle> spriteFrames;
         int currentFrame = 0;
         int totalFrames;
-        double timePerFrame = 0.25; // Adjustable data
+        double timePerFrame = 0.3; // Adjustable data
         double timeElapsed = 0;
         int index;
         private int width;
@@ -22,7 +22,7 @@ namespace LegendOfZelda
         {
             itemTexture = texture;
             index = sIndex;
-            spriteFrames = SpriteItemData2.GetRectangleData(index);
+            spriteFrames = SpriteItemData.GetRectangleData(index);
             totalFrames = spriteFrames.Count;
         }
